Get everything working
