+++ .github/workflows/ci.yml
@@ -134,6 +134,11 @@ jobs:
           node-version: ${{ matrix.node_version }}
           cache: 'pnpm'
 
+      - name: Use Deno
+        uses: denoland/setup-deno@v1
+        with:
+          deno-version: v1.19.3
+
       - name: Download Build Artifacts
         uses: actions/download-artifact@v3
