+++ packages/astro/package.json
@@ -107,6 +107,7 @@
     "mime": "^3.0.0",
     "ora": "^6.1.0",
     "parse5": "^6.0.1",
+    "path-browserify": "^1.0.1",
     "path-to-regexp": "^6.2.0",
     "postcss": "^8.4.12",
     "postcss-load-config": "^3.1.3",
@@ -148,6 +149,7 @@
     "@types/mime": "^2.0.3",
     "@types/mocha": "^9.1.0",
     "@types/parse5": "^6.0.3",
+    "@types/path-browserify": "^1.0.0",
     "@types/prettier": "^2.4.4",
     "@types/resolve": "^1.20.1",
     "@types/rimraf": "^3.0.2",
