+++ packages/astro/src/core/logger/core.ts
@@ -1,0 +1,139 @@
+import type { AstroConfig } from '../../@types/astro';
+import { bold, dim  } from 'kleur/colors';
+import stringWidth from 'string-width';
+
+interface LogWritable<T> {
+	write: (chunk: T) => boolean;
+}
+
+export type LoggerLevel = 'debug' | 'info' | 'warn' | 'error' | 'silent'; // same as Pino
+export type LoggerEvent = 'info' | 'warn' | 'error';
+
+export interface LogOptions {
+	dest: LogWritable<LogMessage>;
+	level: LoggerLevel;
+}
+
+function getLoggerLocale(): string {
+	const defaultLocale = 'en-US';
+	if (process.env.LANG) {
+		const extractedLocale = process.env.LANG.split('.')[0].replace(/_/g, '-');
+		// Check if language code is atleast two characters long (ie. en, es).
+		// NOTE: if "c" locale is encountered, the default locale will be returned.
+		if (extractedLocale.length < 2) return defaultLocale;
+		else return extractedLocale.substring(0, 5);
+	} else return defaultLocale;
+}
+
+export const dateTimeFormat = new Intl.DateTimeFormat(getLoggerLocale(), {
+	hour: '2-digit',
+	minute: '2-digit',
+	second: '2-digit',
+});
+
+export interface LogMessage {
+	type: string | null;
+	level: LoggerLevel;
+	message: string;
+	args: Array<any>;
+}
+
+export const levels: Record<LoggerLevel, number> = {
+	debug: 20,
+	info: 30,
+	warn: 40,
+	error: 50,
+	silent: 90,
+};
+
+/** Full logging API */
+export function log(opts: LogOptions, level: LoggerLevel, type: string | null, ...args: Array<any>) {
+	const logLevel = opts.level;
+	const dest = opts.dest;
+	const event: LogMessage = {
+		type,
+		level,
+		args,
+		message: '',
+	};
+
+	// test if this level is enabled or not
+	if (levels[logLevel] > levels[level]) {
+		return; // do nothing
+	}
+
+	dest.write(event);
+}
+
+/** Emit a user-facing message. Useful for UI and other console messages. */
+export function info(opts: LogOptions, type: string | null, ...messages: Array<any>) {
+	return log(opts, 'info', type, ...messages);
+}
+
+/** Emit a warning message. Useful for high-priority messages that aren't necessarily errors. */
+export function warn(opts: LogOptions, type: string | null, ...messages: Array<any>) {
+	return log(opts, 'warn', type, ...messages);
+}
+
+/** Emit a error message, Useful when Astro can't recover from some error. */
+export function error(opts: LogOptions, type: string | null, ...messages: Array<any>) {
+	return log(opts, 'error', type, ...messages);
+}
+
+type LogFn = typeof info | typeof warn | typeof error;
+
+export function table(opts: LogOptions, columns: number[]) {
+	return function logTable(logFn: LogFn, ...input: Array<any>) {
+		const messages = columns.map((len, i) => padStr(input[i].toString(), len));
+		logFn(opts, null, ...messages);
+	};
+}
+
+export function debug(...args: any[]) {
+	if('_astroGlobalDebug' in globalThis) {
+		(globalThis as any)._astroGlobalDebug(...args);
+	}
+}
+
+function padStr(str: string, len: number) {
+	const strLen = stringWidth(str);
+	if (strLen > len) {
+		return str.substring(0, len - 3) + '...';
+	}
+	const spaces = Array.from({ length: len - strLen }, () => ' ').join('');
+	return str + spaces;
+}
+
+export let defaultLogLevel: LoggerLevel;
+if(typeof process !== 'undefined') {
+	if (process.argv.includes('--verbose')) {
+		defaultLogLevel = 'debug';
+	} else if (process.argv.includes('--silent')) {
+		defaultLogLevel = 'silent';
+	} else {
+		defaultLogLevel = 'info';
+	}
+} else {
+	defaultLogLevel = 'info';
+}
+
+/** Print out a timer message for debug() */
+export function timerMessage(message: string, startTime: number = Date.now()) {
+	let timeDiff = Date.now() - startTime;
+	let timeDisplay = timeDiff < 750 ? `${Math.round(timeDiff)}ms` : `${(timeDiff / 1000).toFixed(1)}s`;
+	return `${message}   ${dim(timeDisplay)}`;
+}
+
+/**
+ * A warning that SSR is experimental. Remove when we can.
+ */
+export function warnIfUsingExperimentalSSR(opts: LogOptions, config: AstroConfig) {
+	if (config._ctx.adapter?.serverEntrypoint) {
+		warn(
+			opts,
+			'warning',
+			bold(`Warning:`),
+			` SSR support is still experimental and subject to API changes. If using in production pin your dependencies to prevent accidental breakage.`
+		);
+	}
+}
