+++ packages/astro/src/core/render/route-cache.ts
@@ -1,5 +1,5 @@
 import type { ComponentInstance, GetStaticPathsItem, GetStaticPathsResult, GetStaticPathsResultKeyed, Params, RouteData, RSS } from '../../@types/astro';
-import { LogOptions, warn, debug } from '../logger.js';
+import { LogOptions, warn, debug } from '../logger/core.js';
 
 import { generatePaginateFunction } from './paginate.js';
 import { validateGetStaticPathsModule, validateGetStaticPathsResult } from '../routing/validation.js';
