+++ packages/astro/src/core/routing/manifest/create.ts
@@ -1,12 +1,12 @@
 import type { AstroConfig, ManifestData, RouteData } from '../../../@types/astro';
-import type { LogOptions } from '../../logger';
+import type { LogOptions } from '../../logger/core';
 
 import fs from 'fs';
 import path from 'path';
 import { compile } from 'path-to-regexp';
 import slash from 'slash';
 import { fileURLToPath } from 'url';
-import { warn } from '../../logger.js';
+import { warn } from '../../logger/core.js';
 
 interface Part {
 	content: string;
