+++ packages/integrations/deno/readme.md
@@ -1,53 +1,66 @@
-# @astrojs/node
+# @astrojs/deno
 
-An experimental static-side rendering adapter for use with Node.js servers.
+A server-side rendering adapter for use with Deno targets. Write your code in Astro/Node and deploy to Deno servers.
 
 In your astro.config.mjs use:
 
 ```js
-import nodejs from '@astrojs/node';
+import { defineConfig } from 'astro/config';
+import deno from '@astrojs/deno';
 
-export default {
-  adapter: nodejs()
-}
+export default defineConfig({
+  adapter: deno()
+});
 ```
 
-After performing a build there will be a `dist/server/entry.mjs` module that works like a middleware function. You can use with any framework that supports the Node `request` and `response` objects. For example, with Express you can do:
+After performing a build there will be a `dist/server/entry.mjs` module. You can start a server simply by importing this module:
 
 ```js
-import express from 'express';
-import { handler as ssrHandler } from './dist/server/entry.mjs';
-
-const app = express();
-app.use(ssrHandler);
-
-app.listen(8080);
+import './dist/entry.mjs';
 ```
 
-# Using `http`
+## API
 
-This adapter does not require you use Express and can work with even the `http` and `https` modules. The adapter does following the Expression convention of calling a function when either
+### Adapter options
 
-- A route is not found for the request.
-- There was an error rendering.
-
-You can use these to implement your own 404 behavior like so:
+This adapter automatically starts a server when it is imported. You can configure this through options:
 
 ```js
-import http from 'http';
-import { handler as ssrHandler } from './dist/server/entry.mjs';
+import { defineConfig } from 'astro/config';
+import deno from '@astrojs/deno';
 
-http.createServer(function(req, res) {
-  ssrHandler(req, res, err => {
-    if(err) {
-      res.writeHead(500);
-      res.end(err.toString());
-    } else {
-      // Serve your static assets here maybe?
-      // 404?
-      res.writeHead(404);
-      res.end();
-    }
-  });
-}).listen(8080);
+export default defineConfig({
+  adapter: deno({
+    start: false
+  })
+});
+```
+
+If disabling start you need to write your own web server and use `handle` to render requests:
+
+```ts
+import { serve } from "https://deno.land/std@0.132.0/http/server.ts";
+import { handle } from './dist/entry.mjs';
+
+serve((req: Request) => {
+  // Check the request, maybe do static file handling here.
+
+  return handle(req);
+});
+```
+
+----
+
+You an also pass in a port/hostname to use:
+
+```js
+import { defineConfig } from 'astro/config';
+import deno from '@astrojs/deno';
+
+export default defineConfig({
+  adapter: deno({
+    port: 8081,
+    hostname: 'myhost'
+  })
+});
 ```
