+++ packages/integrations/deno/test/deps.js
@@ -1,0 +1,2 @@
+export * from 'https://deno.land/std@0.110.0/path/mod.ts';
+export * from 'https://deno.land/std@0.132.0/testing/asserts.ts';
