+++ packages/integrations/deno/test/fixtures/basics/astro.config.mjs
@@ -1,0 +1,6 @@
+import { defineConfig } from 'astro/config';
+import deno from '@astrojs/deno';
+
+export default defineConfig({
+	adapter: deno()
+})
