+++ packages/integrations/deno/test/fixtures/basics/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@test/deno-astro-basic",
+  "version": "0.0.0",
+  "private": true,
+  "dependencies": {
+    "astro": "workspace:*",
+    "@astrojs/deno": "workspace:*"
+  }
+}
