+++ packages/integrations/deno/test/fixtures/basics/src/pages/index.astro
@@ -1,0 +1,11 @@
+---
+
+---
+<html>
+<head>
+	<title>Basic App on Deno</title>
+</head>
+<body>
+	<h1>Basic App on Deno</h1>
+</body>
+</html>
